Brought back the `useReorderColumns` and `useResizableColumns` hooks and used them in the `useColumns` hook
