+++ src/components/Table/hooks/useColumns.js
@@ -1,9 +1,7 @@
-import { useMemo } from "react";
+import { equals, props } from "ramda";
 
-import { isPresent, noop } from "neetocist";
-import { equals, move, props } from "ramda";
-
-import SortIcon from "../components/SortIcon";
+import useReorderColumns from "./useReorderColumns";
+import useResizableColumns from "./useResizableColumns";
 
 const useColumns = ({
   columns,
@@ -18,26 +16,13 @@ const useColumns = ({
   onTableChange,
   tableOnChangeProps,
 }) => {
-  const isColumnFixed = column => !!column.fixed;
-
-  const dragProps = isReorderEnabled && {
-    onDragEnd: (fromIndex, toIndex) => {
-      let from = fromIndex;
-      let to = toIndex;
-      if (rowSelection) {
-        from = fromIndex - 1;
-        to = toIndex - 1;
-      }
-
-      if (isColumnFixed(columns[from]) || isColumnFixed(columns[to])) return;
-      const newColumns = move(from, to, columns);
-      setColumns(newColumns);
-      onColumnUpdate(newColumns);
-    },
-    nodeSelector: "th",
-    handleSelector: ".drag-handler",
-    ignoreSelector: ".react-resizable-handle",
-  };
+  const { dragProps } = useReorderColumns({
+    isEnabled: isReorderEnabled,
+    columns,
+    setColumns,
+    onColumnUpdate,
+    rowSelection,
+  });
 
   const handleSort = sorter => {
     let newSortedInfo = { ...sorter };
@@ -62,44 +47,16 @@ const useColumns = ({
     );
   };
 
-  const handleResize =
-    index =>
-    (_, { size }) => {
-      const nextColumns = [...columns];
-      nextColumns[index] = { ...nextColumns[index], width: size.width };
-      setColumns(nextColumns);
-    };
-
-  const computedColumnsData = useMemo(
-    () =>
-      columns.map((col, index) => {
-        const modifiedColumn = {
-          ...col,
-          onHeaderCell: column => ({
-            width: column.width,
-            onResize: isResizeEnabled ? handleResize(index) : noop,
-            onResizeStop: () =>
-              isResizeEnabled ? onColumnUpdate(columns) : noop,
-            isSortable: isPresent(col.sorter),
-            onSort: handleSort,
-            sortedInfo,
-            onColumnHide,
-            isHidable: col.isHidable,
-            column: col,
-          }),
-          sortIcon: SortIcon,
-          sortOrder:
-            sortedInfo.field === col.dataIndex ? sortedInfo.order : null,
-        };
-
-        if (!col.ellipsis) {
-          modifiedColumn.ellipsis = true;
-        }
-
-        return modifiedColumn;
-      }),
-    [columns, sortedInfo, tableOnChangeProps]
-  );
+  const { columns: computedColumnsData } = useResizableColumns({
+    columns,
+    setColumns,
+    isEnabled: isResizeEnabled,
+    onColumnUpdate,
+    handleSort,
+    sortedInfo,
+    onColumnHide,
+    tableOnChangeProps,
+  });
 
   return { dragProps, columns: computedColumnsData };
 };
