+++ src/components/Table/hooks/useResizableColumns.js
@@ -1,0 +1,58 @@
+import { useMemo } from "react";
+
+import { isPresent, noop } from "neetocist";
+
+import SortIcon from "../components/SortIcon";
+
+const useResizableColumns = ({
+  columns,
+  setColumns,
+  isEnabled,
+  onColumnUpdate,
+  handleSort,
+  sortedInfo,
+  onColumnHide,
+  tableOnChangeProps,
+}) => {
+  const handleResize =
+    index =>
+    (_, { size }) => {
+      const nextColumns = [...columns];
+      nextColumns[index] = { ...nextColumns[index], width: size.width };
+      setColumns(nextColumns);
+    };
+
+  const computedColumnsData = useMemo(
+    () =>
+      columns.map((col, index) => {
+        const modifiedColumn = {
+          ...col,
+          onHeaderCell: column => ({
+            width: column.width,
+            onResize: isEnabled ? handleResize(index) : noop,
+            onResizeStop: () => (isEnabled ? onColumnUpdate(columns) : noop),
+            isSortable: isPresent(col.sorter),
+            onSort: handleSort,
+            sortedInfo,
+            onColumnHide,
+            isHidable: col.isHidable,
+            column: col,
+          }),
+          sortIcon: SortIcon,
+          sortOrder:
+            sortedInfo.field === col.dataIndex ? sortedInfo.order : null,
+        };
+
+        if (!col.ellipsis) {
+          modifiedColumn.ellipsis = true;
+        }
+
+        return modifiedColumn;
+      }),
+    [columns, sortedInfo, tableOnChangeProps]
+  );
+
+  return { columns: computedColumnsData };
+};
+
+export default useResizableColumns;
